Add files via upload
Out of a total of 22,211 employees, we have 20,150 active employees and 2,061 terminated employees. The gender breakdown is as follows: 10,264 male employees, 9,332 female employees, and 554 non-conforming employees, making up 50.8%, 46.4%, and 2.7% of the total employee count, respectively. Additionally, we have 15,135 employees working in our headquarters and 5,016 working remotely. This insight further visualizes chats on Employee by Jobtitle,Department and Race.